chore: deprecate old components
