+++ src/components/BlurFade/BlurFade.tsx
