+++ src/components/CardsWithDottedLinesBackground.tsx
@@ -1,7 +1,7 @@
 import Link, { Target } from '@components/Link';
-import { Container } from './v2/LandingPage/Container/Container';
-import { Text } from './v2/LandingPage/Text/Text';
-import { Button } from './v2/Button/Button';
+import { Container } from './LandingPage/Container/Container';
+import { Text } from './LandingPage/Text/Text';
+import { Button } from './Button/Button';
 
 interface Card {
   title: string;
