+++ src/components/DeployOnFleek.tsx
@@ -1,6 +1,6 @@
 import { down } from '@utils/screens';
-import { Text } from './v2/LandingPage/Text/Text';
-import { Container } from './v2/LandingPage/Container/Container';
+import { Text } from './LandingPage/Text/Text';
+import { Container } from './LandingPage/Container/Container';
 
 const IconList = [
   {
