+++ src/components/ExplainerBlocks.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import { Container } from './v2/LandingPage/Container/Container';
-import { Text } from './v2/LandingPage/Text/Text';
+import { Container } from './LandingPage/Container/Container';
+import { Text } from './LandingPage/Text/Text';
 
 const List = [
   {
