+++ src/components/LandingPage/Partners/Partners.tsx
