+++ src/components/PricingCard.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Button, buttonVariants } from './v2/Button/Button';
+import { Button, buttonVariants } from './Button/Button';
 import type { VariantProps } from 'class-variance-authority';
 
 export type Props = {
