+++ src/pages/index.astro
@@ -1,15 +1,15 @@
 ---
 import LandingPageHtml from '@layouts/LandingPage.astro';
-import { Navbar } from '@components/v2/Navbar/Navbar';
-import { Hero } from '@components/v2/LandingPage/Hero/Hero';
-import { Partners } from '@components/v2/LandingPage/Partners/Partners';
+import { Navbar } from '@components/Navbar/Navbar';
+import { Hero } from '@components/LandingPage/Hero/Hero';
+import { Partners } from '@components/LandingPage/Partners/Partners';
 import DeployOnFleek from '@components/DeployOnFleek';
 import ImpressUsers from '@components/ImpressUsers';
 import BuildUseCases from '@components/BuildUseCases';
 import ExplainerBlocks from '@components/ExplainerBlocks';
 import PricingLanding from '@components/PricingLanding';
 import ReadyToLive from '@components/ReadyToLive';
-import Footer from '@components/v2/Footer/Footer';
+import Footer from '@components/Footer/Footer';
 ---
 
 <LandingPageHtml>
